Juggle with some of pytest's internals to get access to all fixtures. This is to address the following case:
  fixture1<scope=package>
  fixture2<scope=package, autouse=true, uses=fixture1>

  test1<uses=fixture2>

A failure happens in fixture2. When pytest_runtest_makereport gets called with item==test1, item.funcargs doesn't show ANY of the package scoped fixtures (I assume because the package scope failed to setup). But we need access to fixture1 result, which didn't fail, in order to get the interfaces and collect logs/screenshots, etc.